+++ linkerd/core/src/test/scala/io/buoyant/linkerd/RouterTest.scala
@@ -1,7 +1,7 @@
 package io.buoyant.linkerd
 
-import com.twitter.finagle.buoyant.DstBindingFactory
 import com.twitter.finagle.{Dtab, Stack}
+import com.twitter.finagle.buoyant.DstBindingFactory
 import io.buoyant.config.Parser
 import io.buoyant.namer.{ConfiguredNamersInterpreter, InterpreterInitializer, TestInterpreterInitializer, TestInterpreter}
 import io.buoyant.router.RoutingFactory
@@ -11,14 +11,20 @@ import org.scalatest.FunSuite
 
 class RouterTest extends FunSuite with Exceptions {
 
+  def parseConfig(
+    yaml: String,
+    protos: Seq[ProtocolInitializer] = Seq(TestProtocol.Plain, TestProtocol.Fancy),
+    interpreters: Seq[InterpreterInitializer] = Seq(TestInterpreterInitializer)
+  ): RouterConfig =
+    Parser.objectMapper(yaml, Iterable(protos, interpreters)).readValue[RouterConfig](yaml)
+
   def parse(
     yaml: String,
     params: Stack.Params = Stack.Params.empty,
     protos: Seq[ProtocolInitializer] = Seq(TestProtocol.Plain, TestProtocol.Fancy),
     interpreters: Seq[InterpreterInitializer] = Seq(TestInterpreterInitializer)
   ): Router = {
-    val mapper = Parser.objectMapper(yaml, Iterable(protos, interpreters))
-    val cfg = mapper.readValue[RouterConfig](yaml)
+    val cfg = parseConfig(yaml, protos, interpreters)
     val interpreter = cfg.interpreter.newInterpreter(cfg.routerParams)
     cfg.router(params + DstBindingFactory.Namer(interpreter))
   }
@@ -95,5 +101,25 @@ servers:
     val router = parse(yaml, Stack.Params.empty)
     val DstBindingFactory.Namer(interpreter) = router.params[DstBindingFactory.Namer]
     assert(interpreter.isInstanceOf[TestInterpreter])
+  }
+
+  test("with retries") {
+    val yaml =
+      """|protocol: plain
+         |client:
+         |  retries:
+         |    backoff:
+         |      kind: jittered
+         |      minMs: 1
+         |      maxMs: 1000
+         |    budget:
+         |      ttlSecs: 30
+         |      minRetriesPerSec: 3
+         |      percentCanRetry: 0.33
+         |""".stripMargin
+    assert(parseConfig(yaml).client.flatMap(_.retries) == Some(RetriesConfig(
+      Some(JitteredBackoffConfig(Some(1), Some(1000))),
+      Some(RetryBudgetConfig(Some(30), Some(3), Some(0.33)))
+    )))
   }
 }